done until creating the api and used the jwt tokens and also used postman

+++ controllers/api/v1/posts_api.js
@@ -19,33 +19,27 @@ module.exports.index = async function (req, res) {
     });
 }
 
+
 module.exports.destroy = async function (req, res) {
+
     try {
         let post = await Post.findById(req.params.id);
-        // .id means converting the object id into string
-        // if (post.user == req.user.id) {
+        
+        if (post.user == req.user.id) {
             post.remove();
 
             await Comment.deleteMany({ post: req.params.id });
-            
-            // if (req.xhr){
-            //     return res.status(200).json({
-            //         data: {
-            //             post_id: req.params.id
-            //         },
-            //         message: "Post deleted"
-            //     });
-            // }
 
             
             return res.json(200 , {
                 message : "Post and associated comments deleted successfully"
             });
-        
-        // else {
-        //     req.flash('success' , 'You cannot delete this post');
-        //     return res.redirect('back');
-        // }
+        }
+        else {
+            return res.json(401 , {
+                message : "You cannot delete this post"
+            });
+        }
     }
     catch (err) {
         console.log('Error in deleting a post', err);
